python/facebook-image-poster/requirements.txt
requests
Pillow
selenium
PyVirtualDisplay